Major refactor Got rid of Microcode Engine. Kept MicroCode. Consolidated into Mos6502 structure.
